python_misc/pyproject.toml
[project]
name = "worldcube-python"
version = "0.1.0"
requires-python = "==3.12.*"
dependencies = [
    "mediapipe~=0.10.21",
    "opencv-python",
    "spatium==1.5.5",
#    "vedo==2025.5.4"
    "vedo@git+https://github.com/marcomusy/vedo",
    "pykalman==0.10.1",
    "pyserial==3.5"
]
